requirements.txt
airflow-provider-great-expectations==0.2.0
great-expectations==0.15.32
minio